Pequeno ajuste.

+++ software/pc.c.h
@@ -59,7 +59,7 @@ void PC_Tick (void) {
     Serial.print(F(";"));
     Serial.print(G.hits);             // total de golpes
     Serial.print(F(";"));
-    Serial.print(G.pace[0]);          // media simples
+    Serial.print((int)G.pace[0]);     // media simples
     Serial.print(F(";"));
     Serial.println();
 }
